# app/api-docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Payment Validator API - Documentación del Sistema",
    "description": "API de conciliación de pagos y manejo de Webhooks para la pasarela de procesamiento Payment Validator.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor de Desarrollo Local"
    }
  ],
  "paths": {
...
    "/api/webhooks/binance-incoming": {
      "post": {
        "summary": "Recibir transacciones de Binance (Webhook Interno)",
        "description": "Endpoint seguro que recibe los datos de transacciones procesadas por el Binance Worker y las registra como pendientes en la base de datos Neon.",
        "tags": ["Webhooks"],
        "parameters": [
          {
            "in": "header",
            "name": "X-Webhook-Secret",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Secreto compartido entre el Worker y este backend para autorizar la petición."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["binance_id_completo", "monto", "moneda"],
                "properties": {
                  "binance_id_completo": {
                    "type": "string",
                    "description": "El ID completo de la transacción provisto por Binance.",
                    "example": "PAY_1702484920000"
                  },
                  "monto": {
                    "type": "number",
                    "format": "float",
                    "description": "Monto de la transacción.",
                    "example": 50.00
                  },
                  "moneda": {
                    "type": "string",
                    "description": "Criptomoneda usada en Binance.",
                    "example": "USDT"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transacción pre-registrada exitosamente o detectada como duplicado seguro.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Transacción pre-registrada desde Binance"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado (Falta o es incorrecto el X-Webhook-Secret)."
          },
          "500": {
            "description": "Error interno del servidor o de base de datos."
          },
          "503": {
            "description": "Fallo temporal de conexión con la base de datos: la transacción NO se registró y el worker debe reenviarla."
          }
        }
      }
    }
  }
}